requirements.txt
# Web frameworks
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
xgboost>=1.7.6
tensorflow>=2.15.0; platform_system != "Darwin"
tensorflow-macos>=2.15.0; platform_system == "Darwin"
joblib>=1.3.0
scipy>=1.11.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
statsmodels>=0.14.0

# API and data handling
requests>=2.31.0
pydantic>=2.4.0
PyYAML>=6.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
